third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-control-expected.txt

This is a testharness.js-based test.
[FAIL] pointer updates (control test)
  assert_equals: error events on c expected 1 but got 0
Harness: the test ran to completion.
